# ggml : add ALiBi support for ggml_soft_max_ext