ref: https://github.com/ggerganov/llama.cpp/issues/3470

- [x] CPU
- [x] Metal
- [x] CUDA
- [ ] Vulkan (cc @0cc4m)
- [ ] SYCL (cc @abhilash1910)
- [ ] Kompute (cc @cebtenzzre)

Missing backends currently generate compile warnings. For some time, we will fallback to the old implementation of `ggml_alibi` but this will be removed at some point, so support should be added in all backends when possible.

Additional changes:

- Support multi-sequence ALiBi via `KQ_pos` tensor. It provides the token position for each KV cache cell
- Remove F16 CUDA soft_max kernels