@slaren 2024-02-14
Isn't this softmax part of flash attention? Are we giving up on that?