@ggerganov 2024-02-14
`ggml_soft_max_ext` is needed even with `ggml_flash_attention_ext` because not all backends will implement FA. This way they can fallback to `ggml_soft_max_ext` that fuses scaling, masking and ALiBi slope

Regarding the FA branch - I've hit a roadblock with the CUDA implementation, and don't want to merge just Metal support. But once this is resolved, the `ggml_flash_attention_ext` will be extended with a similar ALiBi support