@ggerganov 2024-02-14
Is there significant benefit from the F16 CUDA soft_max implementation?

https://github.com/ggerganov/llama.cpp/blob/a0f8a93bf126b713ca02193ae67e336843d79910/ggml-cuda.cu#L9132-L9148

I think it is currently broken - enable the `mask` in the `test-backend-ops` and build with `LLAMA_CUDA_F16=1`. 

Also, while on this topic, I see we have custom template specifications for a set of sizes in CUDA soft_max:

https://github.com/ggerganov/llama.cpp/blob/a0f8a93bf126b713ca02193ae67e336843d79910/ggml-cuda.cu#L7673-L7706

How much do we gain from these? I assume there is some benefit for prompt processing, because for TG we almost never hit these branches. cc @JohannesGaessler Do you have some numbers handy?