@JohannesGaessler 2024-02-14
>ggml_soft_max_ext is needed even with ggml_flash_attention_ext because not all backends will implement FA. This way they can fallback to ggml_soft_max_ext that fuses scaling, masking and ALiBi slope

Do consider that if you ever want to implement sampling on the GPU softmax will also be needed.

>Regarding the FA branch - I've hit a roadblock with the CUDA implementation, and don't want to merge just Metal support. But once this is resolved, the ggml_flash_attention_ext will be extended with a similar ALiBi support

As I said before, my intention with FA was to let Steward/FSSRepo handle it (with me advising) since long term I would like to have more devs. But recently there hasn't really been progress from his side so maybe I'll just do it myself in the next few weeks.

>Is there significant benefit from the F16 CUDA soft_max implementation?

A quick low sample size test where I enabled FP16 soft max and nothing else via a simple patch:

| GPU        | Model           |     Batch size | Test     |     t/s master |   t/s patch |     Speedup |
| :--------- | :-------------- | -------------: | :------- | -------------: | ----------: | ----------: |
| RTX 3090   | llama 7B Q4_0   |              1 | pp4096   |         110.83 |      111.15 |        1.00 |
| RTX 3090   | llama 7B Q4_0   |              2 | pp4096   |         209.82 |      210.48 |        1.00 |
| RTX 3090   | llama 7B Q4_0   |              4 | pp4096   |         343.02 |      346.16 |        1.01 |
| RTX 3090   | llama 7B Q4_0   |              8 | pp4096   |         454.09 |      457.24 |        1.01 |
| RTX 3090   | llama 7B Q4_0   |             16 | pp4096   |         462.47 |      465.90 |        1.01 |
| RTX 3090   | llama 7B Q4_0   |             32 | pp4096   |         584.72 |      590.74 |        1.01 |
| RTX 3090   | llama 7B Q4_0   |             64 | pp4096   |        1182.61 |     1192.27 |        1.01 |
| RTX 3090   | llama 7B Q4_0   |            128 | pp4096   |        1906.49 |     1914.28 |        1.00 |
| RTX 3090   | llama 7B Q4_0   |            256 | pp4096   |        2619.82 |     2670.42 |        1.02 |
| RTX 3090   | llama 7B Q4_0   |            512 | pp4096   |        3017.57 |     3073.57 |        1.02 |
| RTX 3090   | llama 7B Q4_0   |           1024 | pp4096   |        3334.44 |     3383.46 |        1.01 |
| RTX 3090   | llama 7B Q4_0   |           2048 | pp4096   |        3303.83 |     3317.45 |        1.00 |
| RTX 3090   | llama 7B Q4_0   |           4096 | pp4096   |        3023.11 |     3037.88 |        1.00 |

The FP16 kernel itself is ~1.4x faster than the FP32 kernel.

>I think it is currently broken on master - enable the mask in the test-backend-ops and build with LLAMA_CUDA_F16=1.

This is an issue with how the tests are set up. The worst precision I get for FP16 softmax is NMSE = 4e-6. This is worse than the threshold of 1e-7 but still completely negligible. In the first place, NMSE does not accurately reflect the effects of numerical precision for softmax. NMSE is dominated by the effects on values that are ~0 after softmax but those values are going to be ignored anyways and even large relative differences like 1e-9 <-> 1e-10 on those values are completely meaningless. 

Also notice how NMSE increases with the input size. I didn't analyze the code in detail but what I think is happening is that the number of values that are ~0 after softmax increases. And because the numerical stability for those values is worse worse NMSE increases. But this does not make a practical difference.