@JohannesGaessler 2024-02-14
>How much do we gain from these?

From the templating in particular I think the difference was ~10% in terms of the kernel runtime.