@ggerganov 2024-02-15
On `master` if I make the following patch:

```diff
diff --git a/tests/test-backend-ops.cpp b/tests/test-backend-ops.cpp
index 9af8517d..92adf4fb 100644
--- a/tests/test-backend-ops.cpp
+++ b/tests/test-backend-ops.cpp
@@ -2091,7 +2091,7 @@ static bool test_backend(ggml_backend_t backend, test_mode mode, const char * op
         for (int n = 0; n < 10; ++n) {
             int64_t ne0 = dist_ne0(rng);
             int64_t ne1 = dist_ne1(rng);
-            test_cases.emplace_back(new test_soft_max(GGML_TYPE_F32, {ne0, ne1, 1, 1}));
+            test_cases.emplace_back(new test_soft_max(GGML_TYPE_F32, {ne0, ne1, 1, 1}, 0.1f));
         }
 
         exponent <<= 1;
```

Some of the tests fail with `inf` NMSE:

```
  SOFT_MAX(type=f32,ne=[55701,6,1,1],scale=0.100000,mask=0): [SOFT_MAX] NMSE = 0.000001874 > 0.000000100 FAIL
  SOFT_MAX(type=f32,ne=[57757,27,1,1],scale=0.100000,mask=0): [SOFT_MAX] NMSE = 0.000003197 > 0.000000100 FAIL
  SOFT_MAX(type=f32,ne=[50919,30,1,1],scale=0.100000,mask=0): [SOFT_MAX] NMSE = 0.000002142 > 0.000000100 FAIL
  SOFT_MAX(type=f32,ne=[43571,36,1,1],scale=0.100000,mask=0): [SOFT_MAX] NMSE = 0.000000810 > 0.000000100 FAIL
  SOFT_MAX(type=f32,ne=[37455,9,1,1],scale=0.100000,mask=0): [SOFT_MAX] NMSE = 0.000001378 > 0.000000100 FAIL
  SOFT_MAX(type=f32,ne=[48671,44,1,1],scale=0.100000,mask=0): [SOFT_MAX] NMSE = 0.000001405 > 0.000000100 FAIL
  SOFT_MAX(type=f32,ne=[118835,28,1,1],scale=0.100000,mask=0): [SOFT_MAX] NMSE = inf > 0.000000100 FAIL
  SOFT_MAX(type=f32,ne=[113968,16,1,1],scale=0.100000,mask=0): [SOFT_MAX] NMSE = inf > 0.000000100 FAIL
  SOFT_MAX(type=f32,ne=[74405,27,1,1],scale=0.100000,mask=0): [SOFT_MAX] NMSE = inf > 0.000000100 FAIL
  SOFT_MAX(type=f32,ne=[85901,30,1,1],scale=0.100000,mask=0): [SOFT_MAX] NMSE = inf > 0.000000100 FAIL
  SOFT_MAX(type=f32,ne=[99490,22,1,1],scale=0.100000,mask=0): [SOFT_MAX] NMSE = inf > 0.000000100 FAIL
  SOFT_MAX(type=f32,ne=[82500,19,1,1],scale=0.100000,mask=0): [SOFT_MAX] NMSE = inf > 0.000000100 FAIL
  SOFT_MAX(type=f32,ne=[91293,23,1,1],scale=0.100000,mask=0): [SOFT_MAX] NMSE = inf > 0.000000100 FAIL
  SOFT_MAX(type=f32,ne=[96725,20,1,1],scale=0.100000,mask=0): [SOFT_MAX] NMSE = inf > 0.000000100 FAIL
  SOFT_MAX(type=f32,ne=[83840,4,1,1],scale=0.100000,mask=0): [SOFT_MAX] NMSE = inf > 0.000000100 FAIL
  SOFT_MAX(type=f32,ne=[89768,13,1,1],scale=0.100000,mask=0): [SOFT_MAX] NMSE = inf > 0.000000100 FAIL
```

They don't fail with the default scale factor of `1.0f` (i.e. the NSME is still larger than the threshold, but it's not `inf`).

I'm thinking we should probably drop the F16 variant of soft_max for now - does not seem really worth to me. Long-term when `ggml` supports F16 activations, we can revisit