@JohannesGaessler 2024-02-15
>I'm thinking we should probably drop the F16 variant of soft_max for now - does not seem really worth to me. Long-term when ggml supports F16 activations, we can revisit

Would be fine with me.