@slaren 2024-02-15
This seems to cause a significant performance regression in alibi models: 

| GPU         | Model          | Test   |   t/s master |   t/s gg/refactor-alibi |   Speedup |
|:------------|:---------------|:-------|-------------:|------------------------:|----------:|
| RTX 3090 Ti | refact 1B Q4_0 | pp512  |      9836.03 |                 4157.76 |      0.42 |
| RTX 3090 Ti | refact 1B Q4_0 | tg128  |       211.99 |                  219.80 |      1.04 |

This can probably be fixed, but I don't really understand what is the goal of this change. I doesn't seem to improve performance, and it will require changes in every backend, it will become obsolete after the introduction of flash attention, and in my opinion the code is not really any simpler, it replaces a very simply `ggml_alibi` operation with a much more complex fused operation.