@ggerganov 2024-02-15
I suppose I did something wrong in the CUDA implementation. Here are the results with Metal:

| model                 |     params | backend    | ngl | test       |    master    t/s |     PR       t/s | speedup |
| --------------------- | ---------: | ---------- | --: | ---------- | ---------------: | ---------------: | ------: |
| refact 1B F16         |     1.59 B | Metal      |  99 | pp 512     |  3564.96 ± 57.21 |  4873.12 ± 110.9 | 1.37    |
| refact 1B F16         |     1.59 B | Metal      |  99 | tg 128     |    108.03 ± 0.03 |    113.72 ± 0.06 | 1.05    |
| mpt 7B F16 (guessed)  |     6.86 B | Metal      |  99 | pp 512     |   1307.48 ± 9.86 |   1454.83 ± 2.39 | 1.11    |
| mpt 7B F16 (guessed)  |     6.86 B | Metal      |  99 | tg 128     |     40.10 ± 0.03 |     40.84 ± 0.02 | 1.02    |

The `ggml_alibi()` on `master` still needs a similar update across all backends because it currently assumes the KV cache consists of a single sequence (i.e. the token position is mapped to the index `i` of the KV cell). Basically, all ALiBi models on `master` work only for single-sequence generation.

It has to improve the performance since we avoid extra read/writes of the entire KV cache.

I don't think Flash Attention will be implemented for every backend. At the very least, on the CPU, FA is slower so most likely we will fallback to `ggml_soft_max_ext` in that case. But likely for some of the other backends too.