@slaren 2024-02-15
I also see a signifnicant speedup in the M3 Max as well:
| Model          |   Model Size [GiB] |   Num. of Parameters | Test   |   t/s master |   t/s gg/refactor-alibi |   Speedup |
|:---------------|-------------------:|---------------------:|:-------|-------------:|------------------------:|----------:|
| refact 1B Q4_0 |               0.86 |           1585842176 | pp512  |      2143.80 |                 2778.23 |      1.30 |
| refact 1B Q4_0 |               0.86 |           1585842176 | tg128  |       141.62 |                  148.71 |      1.05 |


I can see that removing an read/write from the entire KV cache can lead to a large speedup, in that case it would be worth it for the performance improvement alone. We should probably figure a way to make the implementation of fused ops optional in the backends before adding more, though.