@slaren 2024-02-15
The issue with CUDA seems to be the double precision `pow`:

| GPU         | Model          | Test   |   t/s master |   t/s gg/refactor-alibi |   Speedup |
|:------------|:---------------|:-------|-------------:|------------------------:|----------:|
| RTX 3090 Ti | refact 1B Q4_0 | pp512  |      9019.70 |                11988.70 |      1.33 |
| RTX 3090 Ti | refact 1B Q4_0 | tg128  |       211.86 |                  224.65 |      1.06 |


```diff
diff --git a/ggml-cuda.cu b/ggml-cuda.cu
index 7285d6de..5b6acf43 100644
--- a/ggml-cuda.cu
+++ b/ggml-cuda.cu
@@ -5974,14 +5974,14 @@ static __global__ void soft_max_f32(const float * x, const float * mask, const f
     // ALiBi
     if (max_bias > 0.0f) {
         const uint32_t n_head_kv   = gridDim.x/nrows_y;
-        const uint32_t n_head_log2 = 1u << (uint32_t) floor(log2((float) n_head_kv));
+        const uint32_t n_head_log2 = 1u << (uint32_t) floor(log2f((float) n_head_kv));

-        const float m0 = pow(2.0f, -(max_bias       ) / n_head_log2);
-        const float m1 = pow(2.0f, -(max_bias / 2.0f) / n_head_log2);
+        const float m0 = powf(2.0f, -(max_bias       ) / n_head_log2);
+        const float m1 = powf(2.0f, -(max_bias / 2.0f) / n_head_log2);

         const int h = rowx/nrows_y; // head index

-        slope = h < n_head_log2 ? pow(m0, h + 1) : pow(m1, 2*(h - n_head_log2) + 1);
+        slope = h < n_head_log2 ? powf(m0, h + 1) : powf(m1, 2*(h - n_head_log2) + 1);
     }

     extern __shared__ float data_soft_max_f32[];
```